+++ modules/image_draw_objects.h
@@ -13,7 +13,7 @@ template<typename E>
 void draw_polygons(matrix_color<E>* img, std::vector<vertex>* vertices, E polyg_color, vertex v1, vertex v2, vertex v3);
 
 template<typename E>
-void draw_polygons_filled(matrix_color<E> *img, std::vector<vertex> *vertices, std::vector<polygon> *polygons, double scaleX, double scaleY, unsigned char* modelColor);
+void draw_polygons_filled(matrix_color<E> *img, std::vector<vertex> *vertices, std::vector<polygon> *polygons, float scaleX, float scaleY, unsigned char* modelColor);
 
 __shared_func__ void calc_triangle_boundaries(matrix_coord& min_coord, matrix_coord& max_coord, vertex& v1, vertex& v2, vertex& v3, matrix& m);
